feat(admin): improve save actions
Close #440

+++ legadilo/feeds/views/feed_categories_admin_views.py
@@ -16,6 +16,7 @@ from django.utils.translation import gettext_lazy as _
 from django.views.decorators.http import require_GET, require_http_methods
 
 from legadilo.feeds.models import FeedCategory
+from legadilo.users.models import User
 from legadilo.users.user_types import AuthenticatedHttpRequest
 
 
@@ -41,70 +42,50 @@ class FeedCategoryForm(forms.ModelForm):
         model = FeedCategory
         fields = ("title",)
 
+    def __init__(self, *args, user: User, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._user = user
 
-@require_http_methods(["GET", "POST"])
-@login_required
-@transaction.atomic()
-def create_feed_category_view(
-    request: AuthenticatedHttpRequest,
-) -> TemplateResponse | HttpResponseRedirect:
-    form = FeedCategoryForm()
-    status = HTTPStatus.OK
-    if request.method == "POST":
-        status, form, feed_category = _create_feed_category(request)
-        if feed_category:
-            return HttpResponseRedirect(
-                reverse("feeds:edit_feed_category", kwargs={"category_id": feed_category.id})
-            )
-
-    return TemplateResponse(
-        request,
-        "feeds/edit_feed_category.html",
-        {
-            "form": form,
-            "breadcrumbs": [
-                (reverse("feeds:feed_category_admin"), _("Feed categories admin")),
-                (reverse("feeds:create_feed_category"), _("Create feed category")),
-            ],
-        },
-        status=status,
-    )
-
-
-def _create_feed_category(
-    request: AuthenticatedHttpRequest,
-) -> tuple[HTTPStatus, FeedCategoryForm, FeedCategory | None]:
-    form = FeedCategoryForm(data=request.POST)
-    if not form.is_valid():
-        return HTTPStatus.BAD_REQUEST, form, None
-
-    try:
-        feed_category = FeedCategory.objects.create(**form.cleaned_data, user=request.user)
-    except IntegrityError:
-        messages.error(
-            request, _("A category with title '%s' already exists.") % form.cleaned_data["title"]
-        )
-        return HTTPStatus.CONFLICT, form, None
-
-    return HTTPStatus.CREATED, form, feed_category
+    @transaction.atomic()
+    def save(self, commit=True):  # noqa: FBT002 Boolean-typed positional argument in function definition
+        self.instance.user = self._user
+        return super().save(commit=commit)
 
 
 @require_http_methods(["GET", "POST"])
 @login_required
 def edit_feed_category_view(
-    request: AuthenticatedHttpRequest, category_id: int
+    request: AuthenticatedHttpRequest, category_id: int | None = None
 ) -> TemplateResponse | HttpResponseRedirect:
-    category = get_object_or_404(FeedCategory, id=category_id, user=request.user)
-    if request.method == "POST" and "delete" in request.POST:
+    category = None
+    if category_id is not None:
+        category = get_object_or_404(FeedCategory, id=category_id, user=request.user)
+
+    if category and request.method == "POST" and "delete" in request.POST:
         category.delete()
         return HttpResponseRedirect(reverse("feeds:feed_category_admin"))
 
-    form = FeedCategoryForm(instance=category)
+    form = FeedCategoryForm(instance=category, user=request.user)
+    status = HTTPStatus.OK
     if request.method == "POST":
-        form = FeedCategoryForm(instance=category, data=request.POST)
-        if form.is_valid():
-            form.save()
+        status, form, category = _handle_category_edition(request, category)
+        if status == HTTPStatus.OK and "save" in request.POST:
+            return HttpResponseRedirect(reverse("feeds:feed_category_admin"))
+        if status == HTTPStatus.OK and "save-add-new" in request.POST:
+            return HttpResponseRedirect(reverse("feeds:create_feed_category"))
+        if category and status == HTTPStatus.OK:
+            return HttpResponseRedirect(
+                reverse("feeds:edit_feed_category", kwargs={"category_id": category.id})
+            )
 
+    last_crumb = (
+        (
+            reverse("feeds:edit_feed_category", kwargs={"category_id": category.id}),
+            _("Edit feed category"),
+        )
+        if category
+        else (reverse("feeds:create_feed_category"), _("Create feed category"))
+    )
     return TemplateResponse(
         request,
         "feeds/edit_feed_category.html",
@@ -113,10 +94,25 @@ def edit_feed_category_view(
             "category": category,
             "breadcrumbs": [
                 (reverse("feeds:feed_category_admin"), _("Feed categories admin")),
-                (
-                    reverse("feeds:edit_feed_category", kwargs={"category_id": category_id}),
-                    _("Edit feed category"),
-                ),
+                last_crumb,
             ],
         },
+        status=status,
     )
+
+
+def _handle_category_edition(
+    request: AuthenticatedHttpRequest, category: FeedCategory | None
+) -> tuple[HTTPStatus, FeedCategoryForm, FeedCategory | None]:
+    form = FeedCategoryForm(request.POST, instance=category, user=request.user)
+    if not form.is_valid():
+        return HTTPStatus.BAD_REQUEST, form, category
+
+    try:
+        category = form.save()
+    except IntegrityError:
+        messages.error(
+            request, _("A category with title '%s' already exists.") % form.cleaned_data["title"]
+        )
+        return HTTPStatus.CONFLICT, form, category
+    return HTTPStatus.OK, form, category
